fixed useEffect errors
re-formatted code to not use dependancies
