+++ src/components/Exchanges/Exchanges.jsx
@@ -41,8 +41,7 @@ function Exchanges() {
     })
     .catch((err) => {
         setError(err);
-      }).finally(() => {
-      });
+      })
 
   });
